Module Ordering_examples.To_int

Examples for Postface.Ordering.to_int

As OCaml's comparison functions do not return an ordering, but an integer :

< 0 to represent "is lower than".

= 0 to represent "is equal to".

> 0 to represent "is greater than".

Having functions for converting orders to integers and vice versa can be useful for sharing existing functions.

Not surprisingly, converting Lt to int returns -1.

# Ordering.(to_int Lt) ;;
- : int = -1

Just as obviously, converting Eq should return 0.

# Ordering.(to_int Eq) ;;
- : int = 0

And the more experienced will easily guess the result of the integer conversion for Gt!

# Ordering.(to_int Gt) ;;
- : int = 1

Predictable, isn't it?