Module Postface

Introduction

While working on YOCaml, XHTMLBoy told us that the interaction between the (deliberately small) standard Preface library and the standard OCaml library could be irritating.

As it is great fun to reinvent the wheel, we decided to build a standard library (yes, another one) that would implement the abstractions proposed by Preface. The aim is that as soon as someone (original contributor or not) finds an interesting and highly reusable function, they add it to the library (with clear documentation and, obviously, unit tests).

API

Data structures

module List : sig ... end

A list is an immutable, finite sequence of elements of the same type.

Data values

module Ordering : sig ... end

Describes a comparison result.

Functions

module Predicate : sig ... end

A function from 'a to bool.

module Equivalence : sig ... end

A function that takes two 'a and returns a bool. It represent the equivalence.

module Comparison : sig ... end

A function for comparison between two values of the same type.