Module Postface.Comparison

A function for comparison between two values of the same type.

Introduction

Types

type 'a t = 'a -> 'a -> Ordering.t

Preface

Total Implementation

module Invariant : Preface.Specs.INVARIANT with type 'a t = 'a t

When the type is contravariant, its module Invariant is identical to contravariant.

module Contravariant : Preface.Specs.CONTRAVARIANT with type 'a t = 'a t

Examples of using Comparisons as a contravariant. This allows a function to be applied to the observed subjects before the comparison is executed.

Note about Decidable and Divisible

Unlike Predicate and Equivalence, the implementations of Divisible and Decidable are less obvious. Indeed, one can consider a predicate to be the composition of two other predicates if both predicates are true (and equivalence is a more general case of predicate), considering the result of a comparison for products or sums is less natural.

Constructing a comparison that acts on products or sums takes advantage of the idempotent nature of the Ordering monoid. This implies that the order of projection in products (in pairs) or in sums (in Either) matters in the final result of the comparison. This makes these two implementations somewhat less useful than Predicate and Equivalence.

module Divisible : Preface.Specs.DIVISIBLE with type 'a t = 'a t

Examples of using Comparisons as a divisible. This allows the construction of more complex comparisons that act on product types.

module Decidable : Preface.Specs.DECIDABLE with type 'a t = 'a t

Examples of using Comparisons as a decidable. This allows the construction of comparisons that vary according to the branches of a sum type.

Partial Implementation

Some implementations need to be fixed, for example, if we monomorphise an comparison, it is possible to treat it as a monoid.

module Monoid (T : Preface.Specs.Types.T0) : Preface.Specs.MONOID with type t = T.t t

API

Helpers for dealing with Stdlib

Since OCaml's Stdlib deals with integer, it could be useful to be able to convert comparisons to returning integer.

val to_stdlib : 'a t -> 'a -> 'a -> int

to_stdlib c converts a Postface comparison function to an OCaml's Stdlib comparison function. See examples

val from_stdlib : ('a -> 'a -> int) -> 'a t

from_stdlib c converts an OCaml's Stdlib comparison function to a Postface comparison function. See examples

Running comparisons

A comparison can be executed. It is only a matter of giving it values.

val run : 'a t -> 'a -> 'a -> Ordering.t

run e x y is equivalent to e x y. It runs the comparison e on the values x and y. See examples

Convert to equivalence

val is_equal : 'a t -> 'a Equivalence.t

is_equal c lift a comparison to an Equivalence on a and b that return true if a is equal to b. false otherwise. See examples

val is_not_equal : 'a t -> 'a Equivalence.t

is_not_equal c lift a comparison to an Equivalence on a and b that return true if a is not equal to b. false otherwise. See examples

val is_greater : 'a t -> 'a Equivalence.t

is_greater c lift a comparison to an Equivalence on a and b that return true if a is greater to b. false otherwise. See examples

val is_lower : 'a t -> 'a Equivalence.t

is_lower c lift a comparison to an Equivalence on a and b that return true if a is lower to b. false otherwise. See examples

Pretty-printers

val pp : Format.formatter -> 'a t -> unit

Infix operators

Infix operators are located in a sub-module Infix which is included in the body of the module.

module Infix : sig ... end

val (>$) : 'b -> 'b t -> 'a t

Infix version of Contravariant.replace.

val ($<) : 'b t -> 'b -> 'a t

Infix flipped version of Contravariant.replace.

val (>$<) : ('a -> 'b) -> 'b t -> 'a t

Infix version of Contravariant.contramap.

val (>&<) : 'b t -> ('a -> 'b) -> 'a t

Infix flipped version of Contravariant.contramap.

val (>*<) : 'a t -> 'b t -> ('a * 'b) t

Infix version of Divisible.divided.

val (>*) : 'a t -> unit t -> 'a t

Discard the value of the second argument.

val (*<) : unit t -> 'a t -> 'a t

Discard the value of the first argument.

val (>|<) : 'a t -> 'b t -> ('a, 'b) Either.t t

Infix version of Decidable.chosen.