Module Postface.Equivalence

A function that takes two 'a and returns a bool. It represent the equivalence.

Introduction

Like for Predicate, dealing with higher order functions leads to the usage of equivalence. Equivalence is just a shortcut for function 'a -> 'a -> bool so they are compatible with the multiple equal function defined into the standard library.

Types

type 'a t = 'a -> 'a -> bool

Preface

Total Implementation

module Invariant : Preface.Specs.INVARIANT with type 'a t = 'a t

When the type is contravariant, its module Invariant is identical to contravariant.

module Contravariant : Preface.Specs.CONTRAVARIANT with type 'a t = 'a t

Examples of using Equivalences as a contravariant. This allows a function to be applied to the observed subjects before the equivalence is executed.

module Divisible : Preface.Specs.DIVISIBLE with type 'a t = 'a t

Examples of using Equivalences as a divisible. This allows the construction of more complex equivalences that act on product types.

module Decidable : Preface.Specs.DECIDABLE with type 'a t = 'a t

Examples of using Equivalences as a decidable. This allows the construction of equivalences that vary according to the branches of a sum type.

Partial Implementation

Some implementations need to be fixed, for example, if we monomorphise an equivalence, it is possible to treat it as a monoid.

module Monoid (T : Preface.Specs.Types.T0) : Preface.Specs.MONOID with type t = T.t t

API

Running equivalences

An equivalence can be executed. It is only a matter of giving it values.

val run : 'a t -> 'a -> 'a -> bool

run e x y is equivalent to e x y. It runs the equivalence e on the values x and y. See examples

Operations on equivalences

Acting on equivalence.

val negate : 'a t -> 'a t

negate e is equivalent to not equivalent. So run (negate e) x y = not (run e x y). See examples

Pretty-printers

val pp : Format.formatter -> 'a t -> unit

Pretty Printers for predicates in the flavour of the ones from de REPL.

Infix operators

Infix operators are located in a sub-module Infix which is included in the body of the module.

module Infix : sig ... end

val (!) : 'a t -> 'a t

Infix version of negate.

val (>$) : 'b -> 'b t -> 'a t

Infix version of Contravariant.replace.

val ($<) : 'b t -> 'b -> 'a t

Infix flipped version of Contravariant.replace.

val (>$<) : ('a -> 'b) -> 'b t -> 'a t

Infix version of Contravariant.contramap.

val (>&<) : 'b t -> ('a -> 'b) -> 'a t

Infix flipped version of Contravariant.contramap.

val (>*<) : 'a t -> 'b t -> ('a * 'b) t

Infix version of Divisible.divided.

val (>*) : 'a t -> unit t -> 'a t

Discard the value of the second argument.

val (*<) : unit t -> 'a t -> 'a t

Discard the value of the first argument.

val (>|<) : 'a t -> 'b t -> ('a, 'b) Either.t t

Infix version of Decidable.chosen.