Module Postface.List

A list is an immutable, finite sequence of elements of the same type.

Introduction

OCaml lists are Singly Linked List and are First class so they can be produced using a bracket and semicolon notation. For example, [5; 6; 7] produces a list with 3 elements, 5, 6 and 7. And can be directly used in pattern matching.

Note about complexity

Lists are not very efficient for indexing and random access. They are best used when you want to add header elements and when you want to traverse them (entirely).

Operation 	Complexity

Add an item in the first position (using cons) 	O(1)

Add an item in the last position 	O(n) where n is the size of the list

List concatenation 	O(n) where n is the size of the first list

Accessing to the i-st element 	O(i)

Finding an element 	O(n) where n is the size of the list

when the length is negative, it should produces an empty lis

Learn more about lists

OCaml.org list tutorial

3.1 List in OCaml Programming: Correct + Efficient + Beautiful (Michael R. Clarkson et al.)

3. Lists and Patterns in Real World OCaml (2nd Edition) (Yaron Minsky, Anil Madhavapeddy et al.)

Types

A list is described by a recursive type which can have only two cases: the empty list (pronounced nil) or a pair of a value and a list of values (pronounced cons).

type 'a t = 'a list =

| []

| :: of 'a * 'a list

Constructors of type list are a syntactic exception, in fact, these writings are equivalent:

[1; 2; 3]

1 :: 2 :: 3 :: []

1 :: [2; 3]

((::) (1, (::) (2, (::) (3, [])))) (but please, who wants to write that!)

Preface

Total Implementation

module Invariant : Preface.Specs.INVARIANT with type 'a t = 'a t

When the type is coavariant, its module Invariant is identical to Functor.

module Functor : Preface.Specs.FUNCTOR with type 'a t = 'a t

Examples of using List as a Functor. This allows function to be applied on each element of a list.

module Applicative : Preface.Specs.Traversable.API_OVER_APPLICATIVE with type 'a t = 'a t

Examples of using List as an Applicative. This allows to write cartesian product in a functional way.

module Alternative : Preface.Specs.ALTERNATIVE with type 'a t = 'a t

module Selective : Preface.Specs.SELECTIVE with type 'a t = 'a t

module Monad : Preface.Specs.Traversable.API_OVER_MONAD with type 'a t = 'a t

module Monad_plus : Preface.Specs.MONAD_PLUS with type 'a t = 'a t

module Foldable : Preface.Specs.FOLDABLE with type 'a t = 'a t

Partial Implementation

Some implementations need to be fixed, for example, if we monomorphise a list, it is possible to treat it as a monoid.

module Monoid (T : Preface.Specs.Types.T0) : Preface.Specs.MONOID with type t = T.t t

API

Creating lists

Functions to create lists, from other structures or from functions.

val empty : 'a t

empty return an empty list. See examples

val cons : 'a -> 'a t -> 'a t

cons x xs is equivalent to x :: xs. It produces a new list with x as a new head and xs as a tail. See examples

val cons_flipped : 'a t -> 'a -> 'a t

cons_flipped xs x is equivalent of cons x xs. It is a conveinent shortcut to avoid excessive point-free style. See examples

val singleton : 'a -> 'a t

singleton x is equivalent to [x]. It produces a list with only x. See examples

val init : length:int -> (int -> 'a) -> 'a t

init ~length:n f is f 0; f 1; f 2; ...; f (n - 1). It produces a list with the left-to-right application of f from 0 to n - 1. If n <= 0 it produces an empty list. See examples

val init' : int -> (int -> 'a) -> 'a t

init without label.

val replicate : length:int -> 'a -> 'a t

replicate ~length:n x produces a list of size n filled by x. If n <= 0 it produces an empty list. replicate ~length x is equivalent to init ~length (fun _ -> x). See examples

val replicate' : int -> 'a -> 'a t

replicate without label.

val unfold : start:'b -> ('b -> ('a * 'b) option) -> 'a t

unfold ~start:x f will produces a list performing f x. If f x return Some (v, new_x), v is added in the list and unfold will continue using new_x as new value to be applied to f. The function is a generator to produce lists of arbitrary size, dynamically. See examples

warning: The function may not terminate if f never returns None.

val unfold' : 'b -> ('b -> ('a * 'b) option) -> 'a t

unfold without label.

val from_seq : 'a Seq.t -> 'a t

from_seq s will produces a list from a given sequence s. See examples

warning: The function may not terminate if s is an inifnite sequence.

val from_string : string -> char t

from_string s will produces a list of char from a given string s. See examples

val from_bytes : Bytes.t -> char t

from_bytes b will produces a list of char from a given bytes sequence b. See examples

val from_array : 'a Array.t -> 'a t

from_array a will produces a list from a given array a. See examples

val range : start:int -> stop:int -> int t

range ~start:x ~stop:y produces a list of integers from x to y (both included). See examples

val range' : int -> int -> int t

range without label.

Fact about length

val is_empty : 'a t Predicate.t

is_empty l return true if the list is empty, false otherwise. See examples

val length : 'a t -> int

length l return the number of element in the list l. See examples

val compare_lengths : 'a t -> 'b t -> Ordering.t

compare_lengths l1 l2 gives the same result of the comparison of the length of l1 and the length of l2 except that the computation stops after reaching the end of the shortest list. See examples

val compare_length_with : 'a t -> int -> Ordering.t

compare_length_with l1 i gives the same result of the comparison of the length of l1 and the integer i except that the computation stops after at most i iterations on the list. See examples

The following functions can be specialised into predicates, which facilitates their composition.

val has_length : int -> 'a t -> bool

has_length n l return true if l has length n, false otherwise. See examples

val is_larger : int -> 'a t -> bool

is_larger n l return true if l has length greater than n, false otherwise. See examples

val is_smaller : int -> 'a t -> bool

is_smaller n l return true if l has length lower than n, false otherwise. See examples

val largest : 'a t -> 'a t -> 'a t

largest l1 l2 returns the largest list. If size are equal, it returns l2. See examples

val smallest : 'a t -> 'a t -> 'a t

smallest l1 l2 returns the smallest list. If size are equal, it returns l2. See examples

val largest_opt : 'a t -> 'a t -> 'a t option

largest_opt l1 l2 returns the largest list wrapped into an option. If size are equal, it returns None. See examples

val smallest_opt : 'a t -> 'a t -> 'a t option

smallest_opt l1 l2 returns the smallest list wrapped into an option. If size are equal, it returns None. See examples

Pretty-printers

val pp : (Format.formatter -> 'a -> unit) -> Format.formatter -> 'a t -> unit

Pretty Printers for lists.

Infix operators

Infix operators are located in a sub-module Infix which is included in the body of the module.

module Infix : sig ... end

val (--) : int -> int -> int t

Infix version of range.

val (<$>) : ('a -> 'b) -> 'a t -> 'b t

Infix version of Functor.map.

val (<&>) : 'a t -> ('a -> 'b) -> 'b t

Flipped and infix version of Functor.map.

val (<$) : 'a -> 'b t -> 'a t

Infix version of Functor.replace.

val ($>) : 'a t -> 'b -> 'b t

Flipped and infix version of Functor.replace.

val (<*>) : ('a -> 'b) t -> 'a t -> 'b t

Infix version of Applicative.apply.

val (<**>) : 'a t -> ('a -> 'b) t -> 'b t

Flipped and infix version of Applicative.apply.

val (*>) : unit t -> 'a t -> 'a t

Discard the value of the first argument.

val (<*) : 'a t -> unit t -> 'a t

Discard the value of the second argument.

val (<|>) : 'a t -> 'a t -> 'a t

Infix version of Alternative.combine or Monad_plus.combine

val (<*?) : ('a, 'b) Either.t t -> ('a -> 'b) t -> 'b t

Infix version of Selective.select.

val (<||>) : bool t -> bool t -> bool t

Infix version of Selective.or_.

val (<&&>) : bool t -> bool t -> bool t

Infix version of Selective.and_.

val (=|<) : ('a -> 'b) -> 'a t -> 'b t

Infix version of Monad.map.

val (>|=) : 'a t -> ('a -> 'b) -> 'b t

Infix flipped version of Monad.map.

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

Infix flipped version of Monad.bind.

val (=<<) : ('a -> 'b t) -> 'a t -> 'b t

Infix version of Monad.bind.

val (>=>) : ('a -> 'b t) -> ('b -> 'c t) -> 'a -> 'c t

Infix version of Monad.compose_left_to_right.

val (<=<) : ('b -> 'c t) -> ('a -> 'b t) -> 'a -> 'c t

Infix version of Monad.compose_right_to_left.

val (>>) : unit t -> 'b t -> 'b t

Sequentially compose two actions, discarding any value produced by the first.

val (<<) : 'a t -> unit t -> 'a t

Sequentially compose two actions, discarding any value produced by the second.

Syntax operators

Syntax operators are located in a sub-module Syntax which is included in the body of the module.

module Syntax : sig ... end

val let+ : 'a t -> ('a -> 'b) -> 'b t

Syntax flipped version of Functor.map.

val and+ : 'a t -> 'b t -> ('a * 'b) t

Syntax version of Applicative.product.

val let* : 'a t -> ('a -> 'b t) -> 'b t

Syntax flipped version of Monad.bind.