Module List.Infix

val (--) : int -> int -> int t

Infix version of range.

val (<$>) : ('a -> 'b) -> 'a t -> 'b t

Infix version of Functor.map.

val (<&>) : 'a t -> ('a -> 'b) -> 'b t

Flipped and infix version of Functor.map.

val (<$) : 'a -> 'b t -> 'a t

Infix version of Functor.replace.

val ($>) : 'a t -> 'b -> 'b t

Flipped and infix version of Functor.replace.

val (<*>) : ('a -> 'b) t -> 'a t -> 'b t

Infix version of Applicative.apply.

val (<**>) : 'a t -> ('a -> 'b) t -> 'b t

Flipped and infix version of Applicative.apply.

val (*>) : unit t -> 'a t -> 'a t

Discard the value of the first argument.

val (<*) : 'a t -> unit t -> 'a t

Discard the value of the second argument.

val (<|>) : 'a t -> 'a t -> 'a t

Infix version of Alternative.combine or Monad_plus.combine

val (<*?) : ('a, 'b) Either.t t -> ('a -> 'b) t -> 'b t

Infix version of Selective.select.

val (<||>) : bool t -> bool t -> bool t

Infix version of Selective.or_.

val (<&&>) : bool t -> bool t -> bool t

Infix version of Selective.and_.

val (=|<) : ('a -> 'b) -> 'a t -> 'b t

Infix version of Monad.map.

val (>|=) : 'a t -> ('a -> 'b) -> 'b t

Infix flipped version of Monad.map.

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

Infix flipped version of Monad.bind.

val (=<<) : ('a -> 'b t) -> 'a t -> 'b t

Infix version of Monad.bind.

val (>=>) : ('a -> 'b t) -> ('b -> 'c t) -> 'a -> 'c t

Infix version of Monad.compose_left_to_right.

val (<=<) : ('b -> 'c t) -> ('a -> 'b t) -> 'a -> 'c t

Infix version of Monad.compose_right_to_left.

val (>>) : unit t -> 'b t -> 'b t

Sequentially compose two actions, discarding any value produced by the first.

val (<<) : 'a t -> unit t -> 'a t

Sequentially compose two actions, discarding any value produced by the second.