Module Postface.Ordering

Describes a comparison result.

Introduction

The Standard OCaml library uses integers to describe the results of comparisons (if the integer is equal to 0, the objects being compared are equal, if the integer is greater than 0, the first object being compared is larger than the second, and if the integer is less than 0, the first object being compared is smaller than the second), Postface took the decision to use a closed enumeration to reduce the domain of comparison functions.

Types

type t =

| Lt	(*
, for lower than.
*)

| Eq	(*
, for equal.
*)

| Gt	(*
, for greater than.
*)

Preface

Total Implementation

module Monoid : Preface.Specs.MONOID with type t = t

See examples of using ordering as a monoid

API

Helpers for dealing with Stdlib

Since OCaml's Stdlib deals with integer, it could be useful to be able to convert ordering values from int and to int.

val to_int : t -> int

to_int ord, converts Ord.t to int. See examples

val from_int : int -> t

from_int ord, converts int to Ord.t. See examples

Equality between Ord.t

val equal : t -> t -> bool

equal ord1 ord2 return true if ord1 is equivalent to ord2, false otherwise. See examples

val compare : t -> t -> t

compare a b comparison between ordering where Lt < Eq < Gt. See examples

Pretty-printers

val pp : Format.formatter -> t -> unit

Infix operators

Infix operators are located in a sub-module Infix which is included in the body of the module.

module Infix : sig ... end

val (<|>) : t -> t -> t

Infix version of Monoid.combine.