Module Postface.Predicate

A function from 'a to bool.

Introduction

When manipulating higher order functions, it is quite common to use predicates to find a value in a structure or to filter it. That is, functions that for any value, return true or false.

Even if the predicates are only aliases, it is possible to simply compose them. Here is a simple example of a predicate on integers that returns true if the number is even, false otherwise.

# let is_even x = (x mod 2) = 0 ;;
val is_even : int -> bool = <fun>

There is no additional ceremony to describe a predicate than simply defining a one-argument function that returns a boolean.

Types

type 'a t = 'a -> bool

Preface

Total Implementation

module Invariant : Preface.Specs.INVARIANT with type 'a t = 'a t

When the type is contravariant, its module Invariant is identical to contravariant.

module Contravariant : Preface.Specs.CONTRAVARIANT with type 'a t = 'a t

Examples of using Predicates as a contravariant. This allows a function to be applied to the observed subject before the predicate is executed.

module Divisible : Preface.Specs.DIVISIBLE with type 'a t = 'a t

Examples of using Predicates as a divisible. This allows the construction of more complex predicates that act on product types.

module Decidable : Preface.Specs.DECIDABLE with type 'a t = 'a t

Examples of using Predicates as a decidable. This allows the construction of predicates that vary according to the branches of a sum type.

API

Running predicates

A predicate can be executed. It is only a matter of giving it a value.

val run : 'a t -> 'a -> bool

run p x is equivalent to p x. It runs the predicate p on the value x. See examples

Predefined predicates

Some predefined predicates that do not depend on the input.

val tautology : 'a t

A tautology is a predicate that always return true no matter what the input. See examples

val contradiction : 'a t

A contradiction is a predicate that always return false no matter what the input. See examples

Acting on and composing predicates

val negate : 'a t -> 'a t

negate p negates the predicate p so run (negate p) x = not (run p x). See negate_examples for examples. See examples

val and_ : 'a t -> 'a t -> 'a t

and_ p1 p2 produces a predicate that apply p1 and p2 on one input so run (and_ p1 p2) x = (p1 x) && (p2 x). See examples

val or_ : 'a t -> 'a t -> 'a t

or_ p1 p2 produces a predicate that apply p1, and if p1 is not satisfied, it applies p2 on one input so run (or_ p1 p2) x = (p1 x) || (p2 x). See examples

Pretty-printers

val pp : Format.formatter -> 'a t -> unit

Pretty Printers for predicates in the flavour of the ones from de REPL.

Infix operators

Infix operators are located in a sub-module Infix which is included in the body of the module.

module Infix : sig ... end

val (!) : 'a t -> 'a t

Infix version of negate.

val (&&) : 'a t -> 'a t -> 'a t

Infix version of and_.

val (||) : 'a t -> 'a t -> 'a t

Infix version of or_.

val (>$) : 'b -> 'b t -> 'a t

Infix version of Contravariant.replace.

val ($<) : 'b t -> 'b -> 'a t

Infix flipped version of Contravariant.replace.

val (>$<) : ('a -> 'b) -> 'b t -> 'a t

Infix version of Contravariant.contramap.

val (>&<) : 'b t -> ('a -> 'b) -> 'a t

Infix flipped version of Contravariant.contramap.

val (>*<) : 'a t -> 'b t -> ('a * 'b) t

Infix version of Divisible.divided.

val (>*) : 'a t -> unit t -> 'a t

Discard the value of the second argument.

val (*<) : unit t -> 'a t -> 'a t

Discard the value of the first argument.

val (>|<) : 'a t -> 'b t -> ('a, 'b) Either.t t

Infix version of Decidable.chosen.