Module Comparison_examples.As_decidable

Examples of using Comparison as a Postface.Comparison.Decidable

Although we have seen how to compose comparison functions to act on products using Divisible, we have also seen that, due to the idempotent nature of the ordering monoid, the result can be surprising and that the order of observation of the product fields varies the result.

As for predicates and for equivalences can construct comparisons that act on sums, but for the same reasons as in Divisible, the result of these comparisons can be surprising.

Let's look at this strangeness with our good old type int_or_float and its projection function into an either:

# type int_or_float = Int of int | Float of float ;;
type int_or_float = Int of int | Float of float

As before, we need to project our value into a either. Let's use our to_either function again:

# let to_either = function Int x -> Either.left x | Float x -> Either.right x ;;
val to_either : int_or_float -> (int, float) Either/2.t = <fun>

Now we can write a comparison that will act on values of this type, broadly in the same way as for predicates or equivalences:

# let cmp_int = from_stdlib Stdlib.Int.compare ;;
val cmp_int : int Comparison.t = <comparison>
# let cmp_float = from_stdlib Stdlib.Float.compare ;;
val cmp_float : float Comparison.t = <comparison>
# let cmp = to_either >$< Stdlib.(cmp_int>|< cmp_float) ;;
val cmp : int_or_float Comparison.t = <comparison>

So far so good, let's try our new comparison function:

# cmp (Int 10) (Int 10) ;;
- : Ordering.t = Eq
# cmp (Float 5.0) (Float 5.0) ;;
- : Ordering.t = Eq
# cmp (Int 1) (Int 2) ;;
- : Ordering.t = Lt

Great, everything seems to work! Let's quickly test something else:

# cmp (Float 3.0) (Int 12) ;;
- : Ordering.t = Gt

For exactly the same reasons that the order of the fields mattered in the case of Divisible, here a Left value will always be smaller than a Right value. It is partly for these reasons that it is better to construct a specific comparison function than to use Decidable, in this way for example:

# let to_float = function Float x -> x | Int x -> Stdlib.float_of_int x ;;
val to_float : int_or_float -> float = <fun>
# let cmp x y = cmp_float (to_float x) (to_float y) ;;
val cmp : int_or_float -> int_or_float -> Ordering.t = <comparison>

Which does not impose a weighting according to the order of projection in either.