Module Comparison_examples.As_divisible

Examples of using Comparison as a Postface.Comparison.Divisible

While we have seen that the implementations of Divisible for predicates and for equivalences allow the construction of more complex predicates and equivalences for product types, respectively, the same is true for the implementation of Divisible for comparisons. But what does this mean in the case of a product?

The association rules may seem a bit strange, but this is because the monoid described by Ordering is an idempotent monoid, which is also sometimes called a find-first. Yeah when the first element of the operation is not the neutral-one, it takes it a reference. In: Using ordering as a Monoid

As a result, the comparison function will depend on the order of comparison. Let us return to our previous examples. As for predicate, let's re-use our type user.

# type user = {first_name: string; last_name: string; age: int; email: string} ;;
type user = {
  first_name : string;
  last_name : string;
  age : int;
  email : string;
}

Our aim is to provide an comparison function between users. It is obviously possible to provide a direct implementation, but it would not take advantage of Divisible. To take advantage of Divisible, the first step is to normalise our user into a product, for this we can reuse the user_to_couple function we used in the predicate example:

# let user_to_couple u = ((u.first_name, u.last_name), u.age), u.email ;;
val user_to_couple : user -> ((string * string) * int) * string = <fun>

Now we just have to compose our comparisons to build a comparison between users! First, let's define some helpers to make the code more digestible.

# let int = from_stdlib Stdlib.Int.compare ;;
val int : int Comparison.t = <comparison>
# let string = from_stdlib Stdlib.String.compare ;;
val string : string Comparison.t = <comparison>

Now all we have to do is project our user into a couple (remember, it's just contramap and the user_to_couple function), and then we can chain our various comparisons functions with >*<:

# let cmp_user = user_to_couple >$< (string >*< string >*< int >*< string) ;;
val cmp_user : user Comparison.t = <comparison>

Fine, let's create some users and check that our function works.

# let mk_user first_name last_name age email = {first_name; last_name; age; email} ;;
val mk_user : string -> string -> int -> string -> user = <fun>

# let xvw = mk_user "Xavier" "Vdw." 10 "xavier@mail.org" ;;
val xvw : user =
  {first_name = "Xavier"; last_name = "Vdw."; age = 10;
   email = "xavier@mail.org"}
# let grm = mk_user "Pierre" "R."  20 "pierre@mail.org" ;;
val grm : user =
  {first_name = "Pierre"; last_name = "R."; age = 20;
   email = "pierre@mail.org"}
# let dpl = mk_user "Didier" "P." 30 "didier@mail.org" ;;
val dpl : user =
  {first_name = "Didier"; last_name = "P."; age = 30;
   email = "didier@mail.org"}

And this is the strange part, it is always the first element that will, in a way, decide the result of the comparison. We'll only deal with the latter if the previously compared element returns Eq:

# cmp_user xvw xvw ;;
- : Ordering.t = Eq
# cmp_user xvw grm ;;
- : Ordering.t = Gt
# cmp_user dpl grm ;;
- : Ordering.t = Lt

For this reason, a lot of care has been taken in the preamble to this example. The order of the fields is important in the final result of the comparison. For example, if I compare two users who have the same name but different ages (No matter what the email address is), it is the age that will determine the result of the comparison:

# cmp_user (mk_user "a" "a" 20 "a") (mk_user "a" "a" 19 "b") ;;
- : Ordering.t = Gt

This is why, unlike predicates and equivalences, which have a trivial compositional strategy (everything must equal true), the composition of comparison functions is probably less useful.