Module Comparison_examples.From_stdlib

Examples for Postface.Comparison.from_stdlib

The standard OCaml library considers a comparison function to be a function that takes two values and returns an integer. If the integer is positive, then the first value is greater than the second. If the integer is negative, then the second value is greater than the first and if the number is zero, then the two values are equivalent.

The from_stdlib function transforms a standard library comparison (which returns an integer) into a Postface comparison (which returns an Ordering.t):

# (from_stdlib Stdlib.Int.compare) 10 12 ;;
- : Ordering.t = Lt
# (from_stdlib Stdlib.Int.compare) 12 10 ;;
- : Ordering.t = Gt
# (from_stdlib Stdlib.Int.compare) 10 10 ;;
- : Ordering.t = Eq

The primary purpose of this function is to allow the use of comparisons described in the stdlib with functions that expect comparisons that return Ordering.t. It is hoped that when the Postface API is rich enough, this function will no longer be useful.