Module Comparison_examples.Is_equal_and_other

Examples for Postface.Comparison.is_equal, Postface.Comparison.is_not_equal, Postface.Comparison.is_greater and Postface.Comparison.is_lower

An equivalence can be a special case of a comparison. It is therefore possible to promote a comparison into equivalence:

# let equal x y = Comparison.is_equal cmp x y ;;
val equal : 'a -> 'a -> bool = <equivalence>

We can use it exactly as an equivalence:

# equal 10 10 ;;
- : bool = true
# equal 11 10 ;;
- : bool = false

And since an equivalence can be negated, there is a function is_not_equal:

# let is_not_equal x y = Comparison.is_not_equal cmp x y ;;
val is_not_equal : 'a -> 'a -> bool = <equivalence>

We can use it exactly as an equivalence (negated):

# is_not_equal 10 10 ;;
- : bool = false
# is_not_equal 10 12 ;;
- : bool = true

And similar functions for the equivalence with Gt and Lt:

# let is_greater x y = Comparison.is_greater cmp x y ;;
val is_greater : 'a -> 'a -> bool = <equivalence>
# let is_lower x y = Comparison.is_lower cmp x y ;;
val is_lower : 'a -> 'a -> bool = <equivalence>

We can use it exactly as an equivalence:

# is_greater 10 2 ;;
- : bool = true
# is_lower 2 10 ;;
- : bool = true