Module Comparison_examples.Run

Examples for Postface.Comparison.run

Since a comparison is just a function that take two arguments, runis not very useful, indeed, we can easli define a comparison for integer, a comparison function returns a value of type Ordering.t:

# let cmp x y = if x > y then Gt else if x < y then Lt else Eq ;;
val cmp : 'a -> 'a -> Ordering.t = <comparison>

It can now be executed, using, for example, the run function:

# Comparison.run cmp 10 10 ;;
- : Ordering.t = Eq
# Comparison.run cmp 1 12 ;;
- : Ordering.t = Lt
# Comparison.run cmp 12 1 ;;
- : Ordering.t = Gt

Or, without the run function by passing values directly to the comparison!

# cmp 10 10 ;;
- : Ordering.t = Eq
# cmp 1 12 ;;
- : Ordering.t = Lt
# cmp 12 1 ;;
- : Ordering.t = Gt