Module Comparison_examples.To_stdlib

Examples for Postface.Comparison.to_stdlib

The standard OCaml library considers a comparison function to be a function that takes two values and returns an integer. If the integer is positive, then the first value is greater than the second. If the integer is negative, then the second value is greater than the first and if the number is zero, then the two values are equivalent.

The to_stdlib function transforms a Postface comparison (which returns an Ordering.t) into a standard library comparison (which returns an integer):

# (to_stdlib cmp) 10 12 ;;
- : int = -1
# (to_stdlib cmp) 12 10 ;;
- : int = 1
# (to_stdlib cmp) 10 10 ;;
- : int = 0

The primary purpose of this function is to allow the use of comparisons described in Postface with functions that expect comparisons that return integers. It is hoped that when the Postface API is rich enough, this function will no longer be useful.