Module Equivalence_examples.As_contravariant

Examples of using Equivalence as a Postface.Equivalence.Contravariant Functor

As for the use of predicates with a contravariant functor, using contramap (or its infixed version >$<) changes the focus of the equivalence application. If we wanted to write a same_size equivalence we could write it this way:

# let same_size l1 l2 = (List.length >$< Stdlib.Int.equal) l1 l2 ;;
val same_size : 'a list -> 'a list -> bool = <equivalence>

# same_size [] [] ;;
- : bool = true
# same_size [1] [2] ;;
- : bool = true
# same_size [1; 2; 3] [1; 2] ;;
- : bool = false

As with predicates, contramap will first apply a function to the observed subject whose result will be applied to the equivalence passed as an argument.

Now that we have seen how to change the focus, we can look at how to compose several equivalences together to build more complex equivalences, with, Divisible.