Module Equivalence_examples.As_decidable

Examples of using Equivalence as a Postface.Equivalence.Decidable

Using Divisible we observed how to compose multiple equivalences to produces a more complex equivalence that act on product types (projected into couples). As for predicates, Decidable allows you to act on sums, i.e. to apply a different equivalence dependending on the branch you are in. Let's reuse our int_or_float:

# type int_or_float = Int of int | Float of float ;;
type int_or_float = Int of int | Float of float

As before, we need to project our value into a either. Let's use our to_either function again:

# let to_either = function Int x -> Either.left x | Float x -> Either.right x ;;
val to_either : int_or_float -> (int, float) Either/2.t = <fun>

Now that we can project an arbitrary sum into the most generic sum (yes, Either is always the most generic sum), we can compose our two equivalences to selectively choose one or the other.

# let equal = to_either >$< (Int.equal >|< Float.equal) ;;
val equal : int_or_float t = <equivalence>

Now that we have composed our equivalences to choose the appropriate one according to the branch we are in, we can test our new function equal:

# equal (Int 10) (Int 10) ;;
- : bool = true
# equal (Int 10) (Int 11) ;;
- : bool = false
# equal (Float 2.5) (Float 2.5) ;;
- : bool = true
# equal (Float 2.0) (Int 2) ;;
- : bool = false

You will note that Int 2 is different from Float 2.0 because it is arbitrarily considered that two resident values in two different branches are de-facto different.

Following the same logic as for products (and for predicates), we can obviously construct a to_either function for sums with more than two branches:

# type more_than_2_branches = A of int | B of float | C of string ;;
type more_than_2_branches = A of int | B of float | C of string

A function to_either of this form can be produced:

let to_either = function
| A x -> Either.(left (left x))
| B x -> Either.(left (right x))
| C x -> Either.right x
;;

And we can write our equal function in this way:

# let equal = to_either >$< (Int.equal >|< Float.equal >|< String.equal) ;;
val equal : more_than_2_branches t = <equivalence>

We can now test our equivalences:

# equal (A 10) (A 10) ;;
- : bool = true
# equal (B 2.0) (C "foo") ;;
- : bool = false

The projection logic is broadly the same as for products.