Module Equivalence_examples.As_divisible

Examples of using Equivalence as a Postface.Equivalence.Divisible

As for predicate, let's re-use our type user.

# type user = {first_name: string; last_name: string; age: int; email: string} ;;
type user = {
  first_name : string;
  last_name : string;
  age : int;
  email : string;
}

Our aim is to provide an equivalence function between users. It is obviously possible to provide a direct implementation, but it would not take advantage of Divisible. To take advantage of Divisible, the first step is to normalise our user into a product, for this we can reuse the user_to_couple function we used in the predicate example:

# let user_to_couple u = ((u.first_name, u.last_name), u.age), u.email ;;
val user_to_couple : user -> ((string * string) * int) * string = <fun>

Now we just have to compose our equivalences to build an equivalence between users! First, let's define some helpers to make the code more digestible.

# let int = Stdlib.Int.equal ;;
val int : int -> int -> bool = <equivalence>
# let string = Stdlib.String.equal ;;
val string : string -> string -> bool = <equivalence>

Now all we have to do is project our user into a couple (remember, it's just contramap and the user_to_couple function), and then we can chain our various equality functions with >*<:

# let eq_user = user_to_couple >$< (string >*< string >*< int >*< string) ;;
val eq_user : user t = <equivalence>

Fine, let's create some users and check that our function works.

# let mk_user first_name last_name age email = {first_name; last_name; age; email} ;;
val mk_user : string -> string -> int -> string -> user = <fun>

# let xvw = mk_user "Xavier" "Vdw." 10 "xavier@mail.org" ;;
val xvw : user =
  {first_name = "Xavier"; last_name = "Vdw."; age = 10;
   email = "xavier@mail.org"}
# let grm = mk_user "Pierre" "R."  20 "pierre@mail.org" ;;
val grm : user =
  {first_name = "Pierre"; last_name = "R."; age = 20;
   email = "pierre@mail.org"}
# let dpl = mk_user "Didier" "P." 30 "didier@mail.org" ;;
val dpl : user =
  {first_name = "Didier"; last_name = "P."; age = 30;
   email = "didier@mail.org"}

Let's try our fresh equivalence. Applying the equivalence between two different users should return false:

# eq_user xvw grm ;;
- : bool = false
# eq_user grm dpl ;;
- : bool = false
# eq_user dpl xvw ;;
- : bool = false

On the other hand, applying the equivalence to the same value should, logically, return true:

# eq_user xvw xvw ;;
- : bool = true
# eq_user grm grm ;;
- : bool = true
# eq_user dpl dpl ;;
- : bool = true

Now that we can construct equivalences, by composition, that act on products, let's observe Decidable which will allow us to construct equivalences that act on sums!