Module Equivalence_examples.Negate

Examples for Postface.Equivalence.negate

The opposite of an equivalence is non-equivalence. The function negate is used to describe that something is not equal. For example, the function Stdlib.Int.equal returns true if two integers are equal, one can easily describe not_equal in terms of equal:

# let not_equal = Equivalence.negate Stdlib.Int.equal ;;
val not_equal : int t = <equivalence>

Which will return true if two numbers are different and false if two numbers are equal:

# not_equal 10 10 ;;
- : bool = false

# not_equal 10 12 ;;
- : bool = true

As for predicates, there is an infix form: !

# (!Stdlib.Int.equal) 10 10 ;;
- : bool = false

# (!Stdlib.Int.equal) 10 12 ;;
- : bool = true