Module Equivalence_examples.Run

Examples for Postface.Equivalence.run

Since an equivalence is just a function that take two arguments, run is not very useful, indeed, we can easily define an equivalence for integer which returns true if two numbers are equals:

# let of_int = Stdlib.Int.equal ;;
val of_int : int -> int -> bool = <equivalence>

It can now be executed, using, for example, the run function:

# Equivalence.run of_int 10 10 ;;
- : bool = true

Or, without the run function by passing values directly to the equivalence!

# of_int 15 10 ;;
- : bool = false