Module List_examples.As_applicative

Examples of using List as an Applicative

Previously, we saw how to apply a function to each element of our list using the Functor functions.

Applicative allows us to construct Cartesian products quite easily. As a reminder, an example of a Cartesian product of two lists would be a list of pairs containing all possible permutations, for example:

# example_of_cartesian_product [1;2;3] ["foo"; "bar"; "baz"] ;;
- : (int * string) list =
[(1, "foo"); (1, "bar"); (1, "baz"); (2, "foo"); (2, "bar"); (2, "baz");
(3, "foo"); (3, "bar"); (3, "baz")]

If we look at the type of the function apply : ('a -> 'b) list -> 'a list -> 'b list, we can see that it looks very similar to map but we need to "lift" our function into a list. For that we can use the function Postface.List.Applicative.pure (or Postface.List.singleton).

# let product = List.Applicative.pure (fun x y -> x, y) ;;
val product : ('_weak2 -> '_weak3 -> '_weak2 * '_weak3) list = [<fun>]

Now we can use apply or rather <*> which is its infix representation:

# let cartesian_product l1 l2 = product <*> l1 <*> l2 ;;
val cartesian_product :
  '_weak4 list -> '_weak5 list -> ('_weak4 * '_weak5) list = <fun>

# cartesian_product [1;2;3] ["foo"; "bar"; "baz"] ;;
- : (int * string) list =
[(1, "foo"); (1, "bar"); (1, "baz"); (2, "foo"); (2, "bar"); (2, "baz");
(3, "foo"); (3, "bar"); (3, "baz")]

In fact, rather than using pure, it is quite common to combine <$> and <*>:

# let cartesian_product l1 l2 = (fun x y -> x, y) <$> l1 <*> l2 ;;
val cartesian_product : 'a list -> 'b list -> ('a * 'b) list = <fun>