Module List_examples.As_functor

Examples of using List as a Functor

The map function is quite well known for lists. It simply allows you to apply a function to each element of a list. This means that if we have a list of A and a function that goes from A to B. Then we can have a list of B using map.

For example, let's say I have a list of integers [1; 2; 3; 4; 5] and I want to convert that list to a list of strings. As I have the function Stdlib.string_of_int, using map, the implementation will be trivial:

# List.Functor.map Stdlib.string_of_int [1;2;3;4;5] ;;
- : string list = ["1"; "2"; "3"; "4"; "5"]

A functor exposes, in addition, an infix operator for map : <$>, which has the same usage:

# Stdlib.string_of_int <$> [1;2;3;4;5] ;;
- : string list = ["1"; "2"; "3"; "4"; "5"]

As well as an operator let+ for which let+ x = y in f x is equivalent to (fun x -> f x) <$> y which allows to write our function in a style close to the direct style:

# let+ x = [1; 2; 3; 4; 5] in Stdlib.string_of_int x ;;
- : string list = ["1"; "2"; "3"; "4"; "5"]

Since we can apply a function to each element in our list, we could describe a replace function that would replace each element in the list with a constant element:

# let replace x l = (fun _ -> x) <$> l ;;
val replace : 'a -> 'b list -> 'a list = <fun>

# replace "OCaml" [1; 2; 3];;
- : string list = ["OCaml"; "OCaml"; "OCaml"]

But fortunately, it is not necessary to describe this function because it already exists (and has two infixed operators):

# List.Functor.replace "OCaml" [1; 2; 3] ;;
- : string list = ["OCaml"; "OCaml"; "OCaml"]

#  "OCaml" <$  [1; 2; 3] ;;
- : string list = ["OCaml"; "OCaml"; "OCaml"]

# [1; 2; 3] $> "OCaml" ;;
- : string list = ["OCaml"; "OCaml"; "OCaml"]

The important point to remember is that using List as a Functor never alters the size of the list. If I have a list of 10 items, after mapping (or replacing, which is a special case of mapping), I will have always 10 items.