Module List_examples.Compare_lengths

Examples for Postface.List.compare_lengths and Postface.List.compare_length_with

Since we have List.length, one might think that we could simply write compare_lengths in this way, if you remember the use of Contravariant on comparisons:

# let cmp_int x y = Comparison.(from_stdlib Stdlib.Int.compare) x y ;;
val cmp_int : int -> int -> Ordering.t = <comparison>
# let cmp_lengths x y = Comparison.(List.length >$< cmp_int) x y;;
val cmp_lengths : 'a list -> 'a list -> Ordering.t = <comparison>

Which will work very well:

# cmp_lengths [] [1; 2; 3] ;;
- : Ordering.t = Lt

The reason why there is a dedicated function is that it is necessary to calculate the sizes of the two lists to execute our previously defined function. However, List.compare_lengths only needs to traverse the smaller of the two lists and offers exactly the same behaviour:

# List.compare_lengths [] [1; 2; 3] ;;
- : Ordering.t = Lt

# List.compare_lengths [1; 2; 3] [] ;;
- : Ordering.t = Gt

# List.compare_lengths [] [] ;;
- : Ordering.t = Eq

# List.compare_lengths [4; 5; 6] [1; 2; 3] ;;
- : Ordering.t = Eq

Also for practical reasons it is possible to directly compare the size of a list with an integer. As with compare_lengths, the function does not have to traverse the entire list. If the list is larger than the integer, the function will return Gt without having to complete the traversal.

# List.compare_length_with [] 2 ;;
- : Ordering.t = Lt

# List.compare_length_with [] 0 ;;
- : Ordering.t = Eq

# List.compare_length_with [1;2;3] 3 ;;
- : Ordering.t = Eq

# List.compare_length_with [1;2;3] 2 ;;
- : Ordering.t = Gt