Module List_examples.Cons

Examples for Postface.List.cons and Postface.List.cons_flipped

cons x xs produces a new list with x as a new head and xs as a tail. This is the "add on top" (or prepend).

# List.cons 10 List.empty ;;
- : int list = [10]

# List.cons "OCaml" ["is"; "the"; "coolest"; "programming"; "language"] ;;
- : string list =
["OCaml"; "is"; "the"; "coolest"; "programming"; "language"]

There is also a version where the parameters are flipped: cons_flipped:

# List.cons_flipped ["is"; "cool"] "OCaml" ;;
- : string list = ["OCaml"; "is"; "cool"]

Note that cons_flipped does not add to the end of the list, it is a flipped version of cons, as its name suggests.