Module List_examples.From_other_values

Examples for Postface.List.from_seq, Postface.List.from_string, Postface.List.from_bytes and Postface.List.from_array

Sometimes it is useful to be able to produce lists based on other data structures. For example, it is possible to construct a list based on a sequence using from_seq:

# let my_seq = Stdlib.List.to_seq [1; 2; 3; 4] ;;
val my_seq : int Seq.t = <fun>
# List.from_seq my_seq ;;
- : int list = [1; 2; 3; 4]

Beware, however, that a sequence may never end. Converting an infinite sequence to a list never ends.

Another common data structure in OCaml is an array. Arrays share some similarities with lists, however, one will prefer to use an array when one wants to intensively use index access, whereas one will prefer to manipulate a list when one wants to, for example, iterate. from_array is used to convert an array into a list:

# List.from_array [|1; 2; 3; 4|] ;;
- : int list = [1; 2; 3; 4]

from_string and from_bytes are used to convert strings and bytes (a mutable form of string) to a list of characters respectively:

# List.from_string "OCaml" ;;
- : char list = ['O'; 'C'; 'a'; 'm'; 'l']

# List.from_bytes (Stdlib.Bytes.of_string "OCaml") ;;
- : char list = ['O'; 'C'; 'a'; 'm'; 'l']