Module List_examples.Init

Examples for Postface.List.init

init allows you to create a list of a given size by applying a function to an index.

For example, if we just return the index, we will produce a list containing integers from 0 to the given size minus 1 :

# List.init ~length:10 (fun x -> x) ;;
- : int list = [0; 1; 2; 3; 4; 5; 6; 7; 8; 9]

If we wanted, for example, to construct a list containing the first 10 positive even numbers, we would simply multiply the index by two:

# List.init ~length:10 (fun x -> x * 2) ;;
- : int list = [0; 2; 4; 6; 8; 10; 12; 14; 16; 18]

With a little trickery, you could also use init to convert a string to a list of characters:

# let str = "OCaml" ;;
val str : string = "OCaml"
# List.init ~length:(String.length str) (String.get str) ;;
- : char list = ['O'; 'C'; 'a'; 'm'; 'l']

The main difference with the init function in the standard library is that init in Postface never fails. If a negative size is requested, init will return an empty list.

# Stdlib.List.init (-5) (fun x -> x) ;;
Exception: Invalid_argument "List.init".

# Postface.List.init ~length:(-5) (fun x -> x) ;;
- : int list = []