Module List_examples.Is_empty

Examples for Postface.List.is_empty

is_empty is a simple predicate that returns true if the input list is empty and false if it is not.

# List.is_empty [] ;;
- : bool = true

# List.is_empty [1; 2; 3] ;;
- : bool = false

Since the function is a predicate, we can easily express is_not_empty in terms of is_empty:

# let is_not_empty = Predicate.negate List.is_empty ;;
val is_not_empty : '_weak1 list Predicate.t = <predicate>

# is_not_empty [] ;;
- : bool = false

# is_not_empty [1; 2; 3] ;;
- : bool = true