Module List_examples.Largest_smallest

Examples for Postface.List.largest, Postface.List.smallest, Postface.List.largest_opt and Postface.List.smallest_opt

largest and smallest give the larger of the two lists and the smaller of the two lists respectively. If the lists have the same size, both functions will return the second list passed as argument:

# List.largest [1; 2; 3] [4; 5] ;;
- : int list = [1; 2; 3]

# List.smallest [1; 2; 3] [4; 5] ;;
- : int list = [4; 5]

# List.largest [1; 2; 3] [-1; -2; -3] ;;
- : int list = [-1; -2; -3]

If you don't want the second list to be returned in case of a tie there are variations suffixed with _opt which wrap the result in an option and return None in case of a size tie:

# List.largest_opt [1; 2; 3] [4; 5] ;;
- : int list option = Some [1; 2; 3]

# List.smallest_opt [1; 2; 3] [4; 5] ;;
- : int list option = Some [4; 5]

# List.largest_opt [1; 2; 3] [-1; -2; -3] ;;
- : int list option = None