Module List_examples.Range

Examples for Postface.List.range

It is often useful to be able to describe intervals between numerical values. It is of course possible to get away with using init but range is relatively easier to use. It takes two integers and produces a list of all the integers between start and stop. (The bounds are both included):

# List.range ~start:(-5) ~stop:5 ;;
- : int list = [-5; -4; -3; -2; -1; 0; 1; 2; 3; 4; 5]

The interesting point is that start must not be smaller than stop. For example:

# List.range ~start:10 ~stop:(-2) ;;
- : int list = [10; 9; 8; 7; 6; 5; 4; 3; 2; 1; 0; -1; -2]

If the stop bound is equal to the start bound, the resulting list will contain only one element:

# List.range ~start:1 ~stop:1 ;;
- : int list = [1]

There is also an infix operator of range:

# 10 -- 20 ;;
- : int list = [10; 11; 12; 13; 14; 15; 16; 17; 18; 19; 20]