Module List_examples.Replicate

Examples for Postface.List.replicate

replicate will create a list of given size filled with a single value.

# List.replicate ~length:10 "OCaml" ;;
- : string list =
["OCaml"; "OCaml"; "OCaml"; "OCaml"; "OCaml"; "OCaml"; "OCaml"; "OCaml";
"OCaml"; "OCaml"]

The function follows the same rules as for Postface.List.init, if the size is negative, the returned list will be empty:

# List.replicate ~length:(-5) 10 ;;
- : int list = []