Module List_examples.Unfold

Examples for Postface.List.unfold

unfold is a kind of list generator. It is given a starting value and a function is applied to it. If this function returns a pair option Some (x, new_start), we add x to the list and reapply the function on new_start. If the function returns None, the list is returned.

For example, we could write init in terms of unfold using this generation function. As long as our index is smaller than the desired size, we return the pair of our function application on the index and increment the index. Once we have reached our bound, we can end the computation by returning None:

# let continue l f i = if i >= l then None else Some (f i, i + 1) ;;
val continue : int -> (int -> 'a) -> int -> ('a * int) option = <fun>

And we can implement init by starting from 0 and applying our continue function at each iteration:

# let init_with_unfold l f = List.unfold ~start:0 (continue l f) ;;
val init_with_unfold : int -> (int -> 'a) -> 'a list = <fun>

This gives us the same behaviour as init:

# init_with_unfold 10 (fun x -> x) ;;
- : int list = [0; 1; 2; 3; 4; 5; 6; 7; 8; 9]

This includes cases where the requested size is less than or equal to 0:

# init_with_unfold (-12) (fun x -> x) ;;
- : int list = []