Module Ordering_examples.As_monoid

Examples of using Ordering as a Postface.Ordering.Monoid

An Ordering forms a monoid, so it has a neutral element and a combination function.

The neutral element is Eq so when we use combine or the infix operator <|>, when Eq is in the operation it does not influence the computation:

# Ordering.Monoid.combine Eq Lt ;;
- : Ordering.t = Lt

# Ordering.Monoid.combine Gt Eq ;;
- : Ordering.t = Gt

The association rules may seem a bit strange, but this is because the monoid described by Ordering is an idempotent monoid, which is also sometimes called a find-first. Yeah when the first element of the operation is not the neutral-one, it takes it a reference.

# Ordering.Monoid.combine Lt Gt ;;
- : Ordering.t = Lt

# Ordering.Monoid.combine Gt Lt ;;
- : Ordering.t = Gt

Knowing this, we can easily speculate on the behaviour of reduce (which will apply combine on all the ordering present in a list, taking as a starting value the neutral element, here Eq), the result will be always the first element which is not Eq (if there is on element that is not Eq, of course).

Here, the result will be Gt, easy!

# Ordering.Monoid.reduce [Gt; Lt; Lt; Eq; Gt] ;;
- : Ordering.t = Gt

And here, the result will be Lt.

# Ordering.Monoid.reduce [Eq; Eq; Lt; Gt; Gt] ;;
- : Ordering.t = Lt

If, on the other hand, we have nothing more than Eq, we are left with an embarrassing Eq at the end.

# Ordering.Monoid.reduce [Eq; Eq] ;;
- : Ordering.t = Eq

That's all we needed to know about Ordering Monoid!