Module Ordering_examples.Equal

Examples for Postface.Ordering.equal

Note that the equal function is used to test whether two orders are the same, they are not particularly related to Eq. The equal function will simply check that two orders have the same value.

All these assertions are therefore true:

# Ordering.(equal Eq Eq) ;;
- : bool = true

# Ordering.(equal Lt Lt) ;;
- : bool = true

# Ordering.(equal Gt Gt) ;;
- : bool = true

But these are false:

# Ordering.(equal Gt Eq) ;;
- : bool = false

# Ordering.(equal Lt Gt) ;;
- : bool = false

# Ordering.(equal Eq Lt) ;;
- : bool = false

Nothing too complicated.