Module Ordering_examples.From_int

Examples for Postface.Ordering.from_int

This function is the opposite of Ordering.to_int. It converts an integer into an ordering. Allowing the use of the Postface functions with the standard library.

If the integer is negative, we will return Lt, logical isn't it!

# Ordering.from_int (-1) ;;
- : Ordering.t = Lt

If on the other hand our integer is zero, then logically the function will return Eq.

# Ordering.from_int 0 ;;
- : Ordering.t = Eq

And without surprises...

# Ordering.from_int 1 ;;
- : Ordering.t = Gt

Look at To_int example to see the function that converts an integer to an ordering.