Module Predicate_examples.And_or

Examples for Postface.Predicate.and_ and Postface.Predicate.or_

As we have seen for negate we can construct predicates. or_ and and_ (or their infixed forms, respectively || and &&) allow to compose several predicates!

For example, let's assume these two predicates:

# let is_smaller_than x y = x > y ;;
val is_smaller_than : 'a -> 'a -> bool = <fun>

# let is_larger_than x y = x < y ;;
val is_larger_than : 'a -> 'a -> bool = <fun>

They are not quite predicates, but if they are specialised, e.g. is_larger_than 10, you do get a predicate. We could create a new predicate that mimics the fact that a value is included in an interval, using and_:

# open Predicate ;;
# let is_between min max = (is_larger_than min) && (is_smaller_than max) ;;
val is_between : 'a -> 'a -> 'a t = <fun>

Now we can test our predicate with, for example, the values 15 and 30:

# is_between 15 30 7 ;;
- : bool = false

Oh great, it doesn't work, which is logical, 7 is not included between 15 and 30!

# is_between 15 30 22 ;;
- : bool = true

This works, logically, because 22 is indeed between 15 and 30. Great.

Now we would like to be able to define a predicate is_not_between. You're going to tell me, nothing could be easier, just do !is_between... and you would be right... but as the purpose of this example is also to show how to use or_, let's write this predicate using it!

# let is_not_between min max = is_smaller_than min || is_larger_than max ;;
val is_not_between : 'a -> 'a -> 'a t = <fun>

Now that we have our second predicate, we can test it with the previous values and we should have the reverse of our previous results:

# is_not_between 15 30 7 ;;
- : bool = true

# is_not_between 15 30 22 ;;
- : bool = false