Module Predicate_examples.As_contravariant

Examples of using Predicates as a Postface.Predicate.Contravariant Functor

By using negate, and_ and or_ we have seen how to compose predicates with each other, however, we were only acting on predicates of the same type. Sometimes we want to promote/lifter a predicate to act on more complex structures.

For example, consider this function which returns a predicate which checks that an integer is equal to the parameter passed to the function:

# let is_equal_to x = Int.equal x;;
val is_equal_to : int -> int -> bool = <fun>

Let's say we want to use it to calculate the size of a list. We could manually define this function, which would return a predicate that would act on a list:

# let has_size s l = is_equal_to s (List.length l) ;;
val has_size : int -> 'a list -> bool = <fun>

It works perfectly, however, there is a systematic way to change the focus (i.e., transform a predicate by allowing it to act on a given prism) using contramap :

# let has_size s = Contravariant.contramap List.length (is_equal_to s) ;;
val has_size : int -> 'a list t = <fun>

which could be described more concisely, using the infix version of contramap as follows:

# open Contravariant ;;
# let has_size s = List.length >$< (is_equal_to s) ;;
val has_size : int -> 'a list t = <fun>

And now we can test it all. In this case, the function will return false because the list has size 4 and not 10.

# has_size 10 [1; 2; 3; 4] ;;
- : bool = false

Let's try it now with a list size of 4 and an expected size of 4, everything should go better!

# has_size 4 [10; 11; 12; 13] ;;
- : bool = true

To summarise, contramap will apply a function to the value to be checked before applying the predicate. In our example, List.length is the function, and we apply it to our function is_equal_to 10 which produces a new predicate that checks that the size of the list is 10.

Although this may seem trivial, it allows us to generalise predicates generically and can be applied to the more complex constructions that we will look at in Afterword. In addition, in the same way that functors can be used as a basis for application pipelines, it is possible to construct more complex predicates, for example using Divisible.