Module Predicate_examples.As_decidable

Examples of using Predicates as a Postface.Predicate.Decidable

Using Divisible, we observed how to apply multiple predicates on multiple members of a value by transforming it into a product (a couple). Decidable allows you to act on sums, i.e. to apply a different predicate depending on the branch you are in. For example, imagine this type of data:

type int_or_float = Int of int | Float of float

As always, it is possible to define a direct implementation that does not use the functions of the Decidable module at all, for example in this way:

let check_int_or_float (p_int : int t) (p_float : float t) = function
| Int x -> p_int x
| Float x -> p_float x

# check_int_or_float ;;
- : int t -> float t -> int_or_float -> bool = <fun>

But in order to select (decide) which predicate to invoke in branches of sums, it is necessary, as with products, to use a generic representation of sums, and it turns out that while the type 'a * 'b represents the most generic product (of at least two elements) possible, ('a, 'b) Either.t represents the most generic sum (of at least two cases, with only one case, like for Divisible, Contravariant is sufficient) possible. We therefore need to provide a function capable of projecting our int_or_float type into an either.

let to_either = function
| Int x -> Either.left x
| Float x -> Either.right x

# to_either ;;
- : int_or_float -> (int, float) Either/2.t = <fun>

And now we can use >|<, the infix operator for chosen so that the correct predicate is passed to the value:

# let check p_int p_float = to_either >$< (p_int >|< p_float) ;;
val check : int t -> float t -> int_or_float t = <fun>

The scheme follows the same form as seen in Divisible, i.e. : projection >$< (p op p2 op p3). First we project our value into its generic representation with >$<, and then we string our predicates with our basic operator (>*< for Divisible and >|< for Decidable). Without further ado, let's test our function with two predicates. We would like to validate only numbers greater than 10:

# let p = check (fun x -> x > 10) (fun x -> x > 10.0) ;;
val p : int_or_float t = <predicate>
# p (Int 11) ;;
- : bool = true
# p (Float 12.0) ;;
- : bool = true
# p (Int 10) ;;
- : bool = false
# p (Float 9.999999) ;;
- : bool = false

In the same way as for products, it is possible to construct types with more than two branches. For example, let's imagine this type:

type more_than_2_branches = A of int | B of float | C of string  

A function to_either of this form can be produced:

let to_either = function
| A x -> Either.(left (left x))
| B x -> Either.(left (right x))
| C x -> Either.right x

For which we can provide this function check.

# let check a b c = to_either >$< (a >|< b >|< c) ;;
val check : int t -> float t -> string t -> more_than_2_branches t = <fun>

As you can see, the projection logic is broadly the same as for products.