Module Predicate_examples.As_divisible

Examples of using Predicates as a Postface.Predicate.Divisible

As we have seen in the exploitation of contravariant functors for predicates, we can apply a predicate only on a fragment of data, we would like to be able to apply several predicates on several fragments of the same value. For example to validate all the fields.

For example, consider this somewhat naive representation of a user:

type user =
{ first_name : string
; last_name : string
; age : int
; email : string
}

We can easily describe some predicates that validate each of the fields. Without much ado, we simply consider that for a first_name or last_name to be valid, it must be at least 2 characters long...

# let check_name name = Stdlib.String.length name >= 2 ;;
val check_name : string -> bool = <predicate>

Then it is assumed, again very naively, that the age must be greater than or equal to zero.

# let check_age age = age >= 0 ;;
val check_age : int -> bool = <predicate>

And for the email address, we'll just assume there's only one "@" in the text, again, security is not our priority in this example:

# let split = Stdlib.String.split_on_char '@' ;;
val split : string -> string list = <fun>
# let check_email email = List.has_length 2 (split email)  ;;
val check_email : string -> bool = <predicate>

We need to be able to compose our predicates to ensure that a value of type user is valid.

# let first_name {first_name; _} =  check_name first_name ;;
val first_name : user -> bool = <predicate>

# let last_name {last_name; _} = check_name last_name ;;
val last_name : user -> bool = <predicate>

# let age {age; _} = check_age age ;;
val age : user -> bool = <predicate>

# let email {email; _} = check_email email ;;
val email : user -> bool = <predicate>

And now we can wrap our predicates together to check that a user is valid!

# let check_user = first_name && last_name && age && email ;;
val check_user : user t = <predicate>

We do have a predicate that can validate a user, for example:

# let mk_user first_name last_name age email = {first_name; last_name; age; email} ;;
val mk_user : string -> string -> int -> string -> user = <fun>

# let a_valid_user = mk_user "Xavier" "Vdw" 32 "xaviervdw@mail.org" ;;
val a_valid_user : user =
  {first_name = "Xavier"; last_name = "Vdw"; age = 32;
   email = "xaviervdw@mail.org"}

# let an_invalid_user = mk_user "Xavier" "Vdw" (-3) "foo" ;;
val an_invalid_user : user =
  {first_name = "Xavier"; last_name = "Vdw"; age = -3; email = "foo"}

Let's test our predicates:

# check_user a_valid_user ;;
- : bool = true
# check_user an_invalid_user ;;
- : bool = false

Everything seems to work, but you have to admit that it's a bit boring... If only we had a function that allowed us to systematise the validation!

This the role of divide (which has type ('a -> 'b * 'c) -> 'b t -> 'c t -> 'a t). What we need is simply the possibility to change a user into a couple, nothing more simple:

let user_to_couple { first_name; last_name; age; email } =
((first_name, last_name), age), email
;;

# user_to_couple ;;
- : user -> ((string * string) * int) * string = <fun>

And yes, a couple allows to represent a kind of heterogeneous list. So a couple of couples is also a couple. Logical!

let check_user_using_divide =
user_to_couple
>$< (check_name >*< check_name >*< check_age >*< check_email)

Let's test our predicates:

# check_user_using_divide a_valid_user ;;
- : bool = true
# check_user_using_divide an_invalid_user ;;
- : bool = false

Concretely, as long as we can correctly represent our type as a pair, we can arbitrarily compose predicates to validate several fragments of a value.

More generally, Divisible allows us to act on product types and the most generic and minimal product (of at least two elements, if the product as only one member, Contravariant is sufficient) is the couple as it allows encoding a more complex variety of product. Hence the need for a conversion function from user to couple. Interestingly, it is possible to process sums, using Decidable!