Module Predicate_examples.Contradiction

Examples for Postface.Predicate.contradiction

A contradiction is a logical function that will always return false, regardless of its input.

# Predicate.contradiction true ;;
- : bool = false

# Predicate.contradiction false ;;
- : bool = false

# Predicate.contradiction "foo" ;;
- : bool = false

# Predicate.contradiction 45 ;;
- : bool = false

Look at Tautology to see a predicate that always return true.