Module Predicate_examples.Negate

Examples for Postface.Predicate.negate

Having a dedicated module for predicates also allows you to act directly on them, the negate function allows, for example, to reverse a predicate.

For example, we could define a predicate is_odd which would be the negation of is_even (and yes, I think we can formally state that any number which is not even is odd).

# let is_odd = Predicate.negate is_even ;;
val is_odd : int Predicate.t = <predicate>

Or by using the infix ! operator provided for this purpose:

# open Predicate ;;
# let is_odd = !is_even ;;
val is_odd : int t = <predicate>

Now we can proudly use our two predicates symmetrically!

# is_even 15 ;;
- : bool = false
# is_odd 15 ;;
- : bool = true

Look at And, and Or to see how to compose predicates.