Module Predicate_examples.Run

Examples for Postface.Predicate.run

Since a predicate is just a function, run is not very useful, indeed, we can easily define a predicate is_even which returns true if a number is even and false if it is odd:

# let is_even x = x mod 2 = 0 ;;
val is_even : int -> bool = <predicate>

It can now be executed, using, for example, the run function:

# Predicate.run is_even 10 ;;
- : bool = true

Or, without the run function by passing the value directly to the predicate!

# is_even 15 ;;
- : bool = false