Module Predicate_examples.Tautology

Examples for Postface.Predicate.tautology

A tautology is a logical function that will always return true, regardless of its input.

# Predicate.tautology false ;;
- : bool = true

# Predicate.tautology true ;;
- : bool = true

# Predicate.tautology "foo" ;;
- : bool = true

# Predicate.tautology 45 ;;
- : bool = true

Look at Contradiction to see a predicate that always return false.